In case encoder mode doesn't work

+++ firmware/Core/Src/encoderMode.c
@@ -6,12 +6,12 @@
 
 
 #include "encoderMode.h"
-#include <tim.h>
+/*#include <tim.h>
 
-float motorSpeed (uint32_t cnt1){
+float motorSpeed (uint16_t cnt1){
 
-	uint32_t cnt2 = __HAL_TIM_GET_COUNTER(&htim2);
-	uint32_t diff = 0;
+	uint16_t cnt2 = __HAL_TIM_GET_COUNTER(&htim2);
+	uint16_t diff = 0;
 
 	if(__HAL_TIM_IS_TIM_COUNTING_DOWN(&htim2)){
 		if(cnt2 < cnt1)
@@ -25,7 +25,8 @@ float motorSpeed (uint32_t cnt1){
 			diff = (65535 - cnt1) + cnt2;
 	}
 
-	uint32_t speed = (diff/8)*60;  // change when sysTimer configured
+	uint16_t speed = (diff/8)*60;  // change when sysTimer configured
 
 	return speed;
 }
+*/
